+++ src/components/App.jsx
@@ -1,24 +1,7 @@
 import { Link, Route, Routes } from 'react-router-dom';
-import Quote from '../routerpages/QuotePage';
-import Calculator from '../routerpages/CalculatorPage';
-
-const DefaultPageContent = () => (
-  <div className="font-semibold">
-    <p className="text-3xl">Welcome to our page!</p>
-    <p className="my-10">
-      Hey there. I&apos;m Ibrohim. I&apos;m a software developer. I can help you build a product,
-      feature, or website. Look through some of my work and experience! If you like what you see and
-      have a project you need coded, don&apos;t hesitate to contact me.
-    </p>
-    <p className="my-10">
-      calculator app, designed to simplify your mathematical tasks effortlessly. With an intuitive
-      interface and advanced functionality. click here to visit{' '}
-      <Link className="text-blue-300 underline" to="/calculator">
-        Calculator{' '}
-      </Link>
-    </p>
-  </div>
-);
+import Quote from '../page/QuotePage';
+import Calculator from '../page/CalculatorPage';
+import HomePage from '../page/HomePage';
 
 const App = () => {
   return (
@@ -43,7 +26,7 @@ const App = () => {
           </nav>
         </div>
         <Routes>
-          <Route path="/" element={<DefaultPageContent />} />
+          <Route path="/" element={<HomePage />} />
           <Route path="/quote" element={<Quote />} />
           <Route path="/calculator" element={<Calculator />} />
         </Routes>
