+++ src/page/QuotePage.jsx
